Update DB
update db after revisi (review needed)

+++ database/migrations/2020_01_15_035038_create_instansis_table.php
@@ -19,7 +19,7 @@ class CreateInstansisTable extends Migration
             $table->string('negara');
             $table->string('kota');
             $table->string('alamat');
-            $table->unsignedBigInteger('id_kategori');
+            $table->unsignedInteger('id_kategori');
            // $table->foreign('id_kategori')->references('id')->on('kategori_instansis');
             $table->string('parent_id');
             $table->string('no_regis_izin');
